% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions-for-perm.R
\name{intervalReturn}
\alias{intervalReturn}
\title{Converts tree breaks to a more useful vector of breaks.}
\usage{
intervalReturn(x = split.cutleft)
}
\arguments{
\item{x}{The splits out of an object created by the `tree` function: `tree$frame$split.cutleft}
}
\value{
A sorted 1D dataframe of breakpoints.
}
\description{
Converts tree breaks to a more useful vector of breaks.
}

